nfc: assorted minor source cleanups from next
- src/isomorphism/vf2.c: fix a doxygen \sa reference to point at
  igraph_count_automorphisms_bliss() (the bliss-specific entry
  that actually implements this code path).
- src/linalg/arpack.c: drop a redundant `(size_t)` cast in an
  index bounds-check.
- src/misc/cocitation.c: pass `IGRAPH_LOOPS` explicitly to
  igraph_degree() instead of the implicit `true`; they are
  numerically equal but the enum name documents intent.
- src/misc/power_law_fit.c: terminate an `IGRAPH_ERRORF` message
  with a period and remove a few stray blank lines.
- src/operators/compose.c: terminate the "Cannot compose directed
  and undirected graph" error message with a period.
- src/operators/products.c: drop a redundant `const` on the
  pass-by-value `root` parameter of igraph_rooted_product().

All changes match `next` verbatim. Library builds cleanly.

+++ src/linalg/arpack.c
@@ -1618,7 +1618,7 @@ const char* igraph_arpack_error_to_string(igraph_arpack_error_t error) {
         : -1;
     const char* message = 0;
 
-    if (index >= 0 && (size_t) index < sizeof(igraph_i_arpack_error_strings) / sizeof(igraph_i_arpack_error_strings[0])) {
+    if (index >= 0 && index < sizeof(igraph_i_arpack_error_strings) / sizeof(igraph_i_arpack_error_strings[0])) {
         message = igraph_i_arpack_error_strings[index];
     }
